requirements.txt
# PyTorch for M1 Mac with MPS (Metal Performance Shaders) support
torch>=2.0.0
torchvision>=0.15.0

# Image processing libraries
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0
opencv-python-headless==4.9.0.80

# Visualization and progress tracking
matplotlib>=3.7.0
tqdm>=4.65.0

# Web interface for testing
streamlit>=1.28.0

# FastAPI backend for Next.js frontend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
